requirements.txt
psycopg2~=2.9.10
pyyaml~=6.0.2
cerberus~=1.3.6
